Move internals of ScheduleCard into ScrollablePeriodList.vue
Mainly, this is for more flexibility of adding things to the card in ScheduleCard. Currently, the slot is at the bottom, but I'm intending to add other things at the top for tvspace, but I don't want it to interfere with the code that expects the slot to be at the bottom.
